+++ changelog.md
@@ -9,6 +9,8 @@
   * Switched `newTBQueue` and `newTBQueueIO` to accept `Natural` as
     size. [#17](https://github.com/haskell/stm/pull/17)
 
+  * Switched `signalTSemN` and `newTSem` to accept `Natural` and `Integer` respectfully.
+
   * Dropped support for `base < 4.8`, due to above change.
 
 ----